\name{SAID-package}
\alias{SAID-package}
\alias{SAID}
\docType{package}
\title{
\packageTitle{SAID}
}
\description{
\packageDescription{SAID}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SAID}
\packageIndices{SAID}
SAIDsampler samples from the posterior distribution of the Synergistic Antagonistic Interaction Detection model.
}
\author{
\packageAuthor{SAID}

Maintainer: \packageMaintainer{SAID}
}
\references{
[to be added]
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
\code{\link{SAIDsampler}}
}
\examples{
#### To be added ####
}
